5.2.1 Поднимет состояние компонента списка карточек на уровень родительского компонента Main, так как нужно знать активную карточку, чтобы правильно отрисовать пины на карте.
